+++ leveldb/util.go
@@ -22,7 +22,7 @@ func shorten(str string) string {
 
 var bunits = [...]string{"", "Ki", "Mi", "Gi", "Ti"}
 
-func shortenb(bytes int) string {
+func shortenb(bytes int64) string {
 	i := 0
 	for ; bytes > 1024 && i < 4; i++ {
 		bytes /= 1024
@@ -30,7 +30,7 @@ func shortenb(bytes int) string {
 	return fmt.Sprintf("%d%sB", bytes, bunits[i])
 }
 
-func sshortenb(bytes int) string {
+func sshortenb(bytes int64) string {
 	if bytes == 0 {
 		return "~"
 	}
